+++ R/ebirdst_compatibility.R
@@ -2,7 +2,7 @@
 #'
 #' @description
 #' Internal functions to facilitate working with both the 2021 and 2022
-#' versions of pkg{ebirdst} despite the significant changes to the API.
+#' versions of \pkg{ebirdst} despite the significant changes to the API.
 #'
 #' @name ebirdst-compatibility
 NULL